fix: move all functions to base service provider
